Reworked testing:
- removed unused test instances
- added images of some test instances
- added testing of internal operators
- added testing of a graph that uses a custom type as node data
- changed type of 'numVisited' variable in topsort alg from unsigned
  to size_t (this change was missed when reworking internal comparisson operators)

+++ SimpleGraphs.hpp
@@ -1047,7 +1047,7 @@ namespace GraphAlgorithms {
 		}
 
 		std::vector<DataType> topologicalOrdering;
-		unsigned numVisited = 0u;
+		size_t numVisited = 0u;
 
 		for (auto& [node, neighbors] : neighborList) {
 			for (auto& [neighbor, weight] : neighbors) {
